Move Layout to rustc_middle
